+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Changed
 
+- When using the nodejs repl type, automatically set the CLJS compile target to :nodejs
+
 # 0.0-51 (2019-09-11 / 4b6a751)
 
 ## Added